refactor: rename config module to chain_spec
The module is the chain's rulebook — ChainSpec, Fork, ChainSpecConfig —
not application configuration; name it after the concept it models and
free `config` for eventual runtime configuration. Pure rename: root
re-exports keep `ChainSpec`/`Fork`/`ChainSpecConfig` paths stable, only
the module path `eth_light_client::config` becomes
`eth_light_client::chain_spec`.
